Do the right shift of the registers with indexes
This removes the 4 indexes in the op array. The formula is this:
i = (OPS_COUNT - j + x) % 4
where i is the final register index, j is the index inside the
block, x is 0, 1, 2, 3 (which corresponds to a, b, c, d) and % 4
is because there are 4 registers
The '% 4' has been replaced by '& 3' which is the same but quicker.