Get des ecb and cbc encryption working
This only has been tested without padding for openssl, this will
need to be done. Also this will have to be tested further with
the base64 option to be sure it does the right thing. There's also
the decryption mode to implement.

More generally, the cmd_des function will have to be tested for
robustness, especially from an I/O standpoint. After all that
the cmdf directory will have to be sorted out.