Make it possible to use parse_rsa_key from stdin
Also add -raw and -passin options to rsautl. -raw is just for
syntaxic compatibility with openssl (I just can copy the command
by changing the name of the program, but it doesnt do anything
beyond being accepted). -passin is because I plan on using the
same function to parse the rsa key in the rsa command so the passin
option will be available anyway (plus it could be useful for
testing purposes).