Get password on des-encrypted keys
This apparently does not work yet. Maybe the pbkdf is not the same
as traditional DES encryption. Try to generate encrypted keys with
ft_ssl and see if it works with the decryption. If it does not
there's either a problem in the decryption function or in the
encrypted key generation. Otherwise the pbkdf is different. Or maybe
just check openssl sources/info, this must be findable.

Also, get_des_password will have to be changed so that it is not
necessary to type the password two times on decryption (and also
maybe add a parameter for the input message).