Fix pbkdf by copying the salt
In a new and 'improved' version of pbkdf the salt local variable
was removed and a ft_memswap was directly performed on des->salt.
This caused 'ft_ssl des' to be broken and an extra ft_memswap in
rsa_print_key for the salt.

So the local variable has been reintsated and the extra ft_memswap
removed. Now both 'ft_ssl des' 'ft_ssl genrsa' work as they should.