Check rsa key with -check option
This needs to be tested further to be sure that we get all types
of errors and that they are detected like in openssl.

What happens if pgcd(key->p, key->q) != 1 || key->p == key->q ?
Also I don't understand why when:
key->p == fe0a24d1 && key->q == 4 && key->e == 0
openssl has an error but not me (d e not congruent to 1).
Maybe it is just that key->e is an invalid value from the get go
so key->d can't be congruent with it ?